+++ src/functions.py
@@ -258,3 +258,19 @@ def auth(authorization, request, check_ip_address = True, allow_application_toke
         return {"error": False, "discordid": discordid, "userid": userid, "name": name, "roles": roles, "application_token": False}
     
     return {"error": True, "descriptor": ml.tr(request, "unauthorized")}
+
+import zlib
+def compress(s):
+    if type(s) == str:
+        s = s.encode()
+    t = zlib.compress(s)
+    t = b64encode(t).decode()
+    return t
+
+def decompress(s):
+    if type(s) == str:
+        s = s.encode()
+    t = b64decode(s)
+    t = zlib.decompress(t)
+    t = t.decode()
+    return t