+++ src/main.py
@@ -19,15 +19,6 @@ drivershub = """    ____       _                         __  __      __
  / /_/ / /  / /| |/ /  __/ /  (__  )  / __  / /_/ / /_/ /
 /_____/_/  /_/ |___/\___/_/  /____/  /_/ /_/\__,_/_.___/ 
                                                          """
-if __name__ == "__main__":
-    from datetime import datetime
-    currentDateTime = datetime.now()
-    date = currentDateTime.date()
-    year = date.strftime("%Y")
-    print(drivershub)
-    print("Drivers Hub: Backend (v1.10.3)")
-    print(f"Copyright (C) {year} CharlesWithC All rights reserved.")
-    print("")
 
 if len(sys.argv) == 1:
     print("Usage: python3 main.py <config.json>")
@@ -50,17 +41,39 @@ if not os.path.exists(config_path):
     exit(1)
 
 if not "HUB_CONFIG_FILE" in os.environ.keys() or os.environ["HUB_CONFIG_FILE"] == "":
-    print(f"Using config: {config_path}")
     os.environ["HUB_CONFIG_FILE"] = config_path
-    print(f"Environment variable HUB_CONFIG_FILE set to {config_path}")
-    print("")
 
-from app import app, config
+from app import app, config, version
 from db import newconn
 import api
 import multilang
 
 if __name__ == "__main__":
+    from datetime import datetime
+    currentDateTime = datetime.now()
+    date = currentDateTime.date()
+    year = date.strftime("%Y")
+    print(drivershub)
+    print(f"Drivers Hub: Backend ({version})")
+    print(f"Copyright (C) {year} CharlesWithC All rights reserved.")
+    print("")
+
+    if os.path.exists(f"./upgrade.py"):
+        print("Detected upgrade")
+        import upgrade
+        try:
+            if upgrade.target_version == version:
+                print("Running upgrade...")
+                upgrade.run()
+                print("Upgrade completed\n")
+            else:
+                print("Version mismatch, aborted\n")
+        except:
+            import traceback
+            traceback.print_exc()
+            print("Upgrade failed due to errors above, main program exited")
+            exit(1)
+
     print(f"Company Name: {config.vtc_name}")
     print(f"Company Abbreviation: {config.vtc_abbr}")
     print("")